Add cargo_helper module to handle header parse
move header parse logical to cargo_helper
