+++ src/main.rs
@@ -5,20 +5,14 @@ extern crate hyper;
 extern crate libsnatch;
 extern crate num_cpus;
 
-use ansi_term::Colour::{Green, Yellow, Red, White};
+use ansi_term::Colour::{Green, Yellow, Red};
 use clap::{App, Arg};
-use hyper::client::Client;
-use hyper::header::{ByteRangeSpec, Headers, Range};
-use libsnatch::authorization::{AuthorizationHeaderFactory, AuthorizationType, GetAuthorizationType};
 use libsnatch::Bytes;
-use libsnatch::client::GetResponse;
-use libsnatch::contentlength::GetContentLength;
 use libsnatch::download::download_chunks;
-use libsnatch::http_version::ValidateHttpVersion;
 use libsnatch::write::OutputFileWriter;
+use libsnatch::util::prompt_user;
+use libsnatch::cargo_helper::get_cargo_info;
 use std::fs::File;
-use std::io;
-use std::io::Write;
 use std::path::Path;
 use std::process::exit;
 
@@ -73,63 +67,6 @@ fn main() {
                  threads);
     }
 
-    // Run Snatch
-    let hyper_client = Client::new();
-
-    // Get the first response from the server
-    let client_response = hyper_client.get_head_response(&url).unwrap();
-
-    print!("# Waiting a response from the remote server... ");
-
-    if !client_response.version.greater_than_http_11() {
-        println!("{}",
-                 Yellow.bold()
-                     .paint("OK (HTTP version <= 1.0 detected)"));
-    } else {
-        println!("{}", Green.bold().paint("OK !"));
-    }
-
-    let auth_type = client_response.headers.get_authorization_type();
-    let auth_header_factory = match auth_type {
-        Some(a_type) => {
-            match a_type {
-                AuthorizationType::Basic => {
-                    println!("{}",
-                             Yellow.bold()
-                                 .paint("[WARNING] The remote content is protected by Basic \
-                                         Auth.\nPlease to enter below your credential \
-                                         informations."));
-                    let username = prompt_user(White.bold(), "Username:");
-                    let password = prompt_user(White.bold(), "Password:");
-                    Some(AuthorizationHeaderFactory::new(AuthorizationType::Basic,
-                                                         username,
-                                                         Some(password)))
-                }
-                _ => {
-                    println!("{}",
-                             Red.bold()
-                                 .paint(format!("[ERROR] The remote content is protected by {} \
-                                                 Authorization, which is not supported!\nYou \
-                                                 can create a new issue to report this problem \
-                                                 in https://github.\
-                                                 com/derniercri/snatch/issues/new",
-                                                a_type)));
-                    exit(1);
-                }
-            }
-        }
-        None => None,
-    };
-
-    let client_response = match auth_header_factory.clone() {
-        Some(header_factory) => {
-            let mut headers = Headers::new();
-            headers.set(header_factory.build_header());
-            hyper_client.get_head_response_using_headers(&url, headers).unwrap()
-        }
-        None => client_response,
-    };
-
     let local_path = Path::new(&file);
 
     if local_path.exists() {
@@ -154,72 +91,20 @@ fn main() {
         }
     }
 
-    let remote_content_length = match client_response.headers.get_content_length() {
-        Some(remote_content_length) => remote_content_length,
-        None => {
-
-            println!("{}",
-                     Yellow.bold()
-                         .paint("[WARNING] Cannot get the remote content length, using an \
-                                 HEADER request."));
-            println!("{}",
-                     Yellow.bold()
-                         .paint("[WARNING] Trying to send an HTTP request, to get the remote \
-                                 content length..."));
-
-            // Trying to force the server to send to us the remote content length
-            let mut custom_http_header = Headers::new();
-            // HTTP header to get all the remote content - if the response is OK, get the
-            // ContentLength information sent back from the server
-            custom_http_header.set(Range::Bytes(vec![ByteRangeSpec::AllFrom(0)]));
-            // Get a response from the server, using the custom HTTP request
-            let client_response =
-                hyper_client.get_http_response_using_headers(&url, custom_http_header).unwrap();
-            // Try again to get the content length - if this one is unknown again, stop the program
-            match client_response.headers.get_content_length() {
-                Some(remote_content_length) => {
-                    println!("{:?}", client_response);
-                    remote_content_length
-                }
-                None => {
-                    println!("{}",
-                             Red.bold()
-                                 .paint("[ERROR] Second attempt has failed."));
-                    exit(1);
-                }
-            }
-        }
-    };
-
+    let cargo_info = get_cargo_info(&url);
     println!("# Remote content length: {:?} MB",
-             (remote_content_length / 1000000) as Bytes);
+             (cargo_info.content_length / 1000000) as Bytes);
 
     let local_file = File::create(local_path).expect("[ERROR] Cannot create a file !");
 
-    local_file.set_len(remote_content_length)
+    local_file.set_len(cargo_info.content_length)
         .expect("[ERROR] Cannot extend file to download size!");
     let out_file = OutputFileWriter::new(local_file);
 
-    download_chunks(remote_content_length,
-                    out_file,
-                    threads as u64,
-                    &url,
-                    auth_header_factory);
+    download_chunks(cargo_info, out_file, threads as u64, &url);
 
     println!("{} Your download is available in {}",
              Green.bold().paint("Done!"),
              local_path.to_str().unwrap());
 
-}
-
-fn prompt_user(style: ansi_term::Style, prompt: &str) -> String {
-    print!("{} ", style.paint(prompt));
-    io::stdout().flush().expect("[ERROR] Couldn't flush stdout!");
-
-    let mut user_input = String::new();
-    io::stdin()
-        .read_line(&mut user_input)
-        .ok()
-        .expect("[ERROR] Couldn't read line!");
-    String::from(user_input.trim())
 }
